pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "discord-py>=2.6.3",
    "flask>=3.1.2",
    "flask-session>=0.8.0",
    "psycopg2-binary>=2.9.10",
    "pypresence>=4.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
